just removed the pages folder and put Login and MainPage pages under src folder as before
