+++ application/front-end/zero-waste/src/App.jsx
@@ -1,12 +1,12 @@
 import { Routes, Route, Navigate } from "react-router-dom";
-import LoginPage from "./pages/Login/LoginPage";
-import SignupPage from "./pages/Login/SignupPage";
-import ProtectedRoute from "./pages/Login/ProtectedRoute";
-import MainPage from './pages/MainPage/MainPage';
-import AdminPanel from  "./Admin/AdminPanel";
-import ChallengePanel from  "./Admin/ChallengePanel";
-import CommentPanel from  "./Admin/CommentPanel";
-import UserPanel from  "./Admin/UserPanel";
+import LoginPage from "./Login/LoginPage";
+import SignupPage from "./Login/SignupPage";
+import ProtectedRoute from "./Login/ProtectedRoute";
+import MainPage from "./MainPage/MainPage";
+import AdminPanel from "./Admin/AdminPanel";
+import ChallengePanel from "./Admin/ChallengePanel";
+import CommentPanel from "./Admin/CommentPanel";
+import UserPanel from "./Admin/UserPanel";
 
 export default function App() {
   return (
@@ -21,7 +21,6 @@ export default function App() {
       <Route path="/login" element={<LoginPage />} />
       <Route path="/signup" element={<SignupPage />} />
       <Route path="*" element={<LoginPage />} />
-      
     </Routes>
   );
 }
