+++ application/front-end/zero-waste/src/MainPage/MainPage.css
